Refactor demo components to dedicated directory structure
- Created demos directory with subdirectories:
  - button demo components in buttons directory
  - style guide components in styles directory
- Added comprehensive documentation with README files and usage examples
- Created consistent directory structure with index.ts exports
- Added backward compatibility re-exports with @deprecated annotations
- Updated COMPONENT-MOVEMENT-PLAN.md to track changes
