+++ src/components/ui/demos/README.md
@@ -1,0 +1,40 @@
+# Demo Components
+
+This directory contains demo and example components used for showcasing and testing UI components.
+
+## Available Components
+
+### Button Demos
+
+Various demonstration components for showcasing button styles:
+
+```tsx
+import { 
+  Button3DDemo,
+  Button3DBorderDemo,
+  Button3DEnhancedDemo, 
+  Button3DPhysicalDemo 
+} from "@/components/ui/demos/buttons";
+
+<Button3DDemo />
+<Button3DBorderDemo />
+<Button3DEnhancedDemo />
+<Button3DPhysicalDemo />
+```
+
+### Style Guides
+
+Components that showcase and document style systems:
+
+```tsx
+import { ButtonStyleGuide } from "@/components/ui/demos/styles";
+
+<ButtonStyleGuide />
+```
+
+## Usage Guidelines
+
+- Use these components for documentation, testing, and design reference
+- Demo components are not intended for production use in application features
+- They provide visual examples of proper component usage and styling options
+- Style guides are useful for maintaining design consistency across the application
