+++ src/components/ui/demos/buttons/index.ts
@@ -1,0 +1,4 @@
+export * from "./button-3d-demo";
+export * from "./button-3d-border-demo";
+export * from "./button-3d-enhanced-demo";
+export * from "./button-3d-physical-demo";
