+++ src/components/ui/demos/index.ts
@@ -1,0 +1,2 @@
+export * from "./buttons";
+export * from "./styles";
